try,catch

+++ controller/LikeController.js
@@ -8,46 +8,73 @@ const addLike = (req, res) => {
   // 좋아요 추가
   const book_id = req.params.id;
 
-  let authorization = ensureAuthorization(req);
+  let authorization = ensureAuthorization(req, res);
 
-  let sql = 'INSERT INTO likes (user_id, liked_book_id) VALUES (?, ?)';
-  let values = [authorization.id, id];
-  conn.query(sql, values, (err, results) => {
-    if (err) {
-      console.log(err);
-      return res.status(StatusCodes.BAD_REQUEST).end();
-    }
+  if (authorization instanceof jwt.TokenExpiredError) {
+    return res.status(StatusCodes.UNAUTHORIZED).json({
+      message: '로그인 세션이 만료되었습니다. 다시 로그인 하세요.'
+    });
+  } else if (authorization instanceof jwt.JsonWebTokenError) {
+    return res.status(StatusCodes.BAD_REQUEST).json({
+      message: '잘못된 토큰입니다.'
+    });
+  } else {
+    let sql = 'INSERT INTO likes (user_id, liked_book_id) VALUES (?, ?)';
+    let values = [authorization.id, id];
+    conn.query(sql, values, (err, results) => {
+      if (err) {
+        console.log(err);
+        return res.status(StatusCodes.BAD_REQUEST).end();
+      }
 
-    return res.status(StatusCodes.OK).json(results);
-  });
+      return res.status(StatusCodes.OK).json(results);
+    });
+  }
 };
 
 const removeLike = (req, res) => {
   // 좋아요 제거(취소)
   const book_id = req.params.id;
 
-  let authorization = ensureAuthorization(req);
+  let authorization = ensureAuthorization(req, res);
 
-  let sql = 'DELETE FROM likes WHERE user_id = ? AND liked_book_id = ?';
-  let values = [authorization.id, book_id];
-  conn.query(sql, values, (err, results) => {
-    if (err) {
-      console.log(err);
-      return res.status(StatusCodes.BAD_REQUEST).end();
-    }
+  if (authorization instanceof jwt.TokenExpiredError) {
+    return res.status(StatusCodes.UNAUTHORIZED).json({
+      message: '로그인 세션이 만료되었습니다. 다시 로그인 하세요.'
+    });
+  } else if (authorization instanceof jwt.JsonWebTokenError) {
+    return res.status(StatusCodes.BAD_REQUEST).json({
+      message: '잘못된 토큰입니다.'
+    });
+  } else {
+    let sql = 'DELETE FROM likes WHERE user_id = ? AND liked_book_id = ?';
+    let values = [authorization.id, book_id];
+    conn.query(sql, values, (err, results) => {
+      if (err) {
+        console.log(err);
+        return res.status(StatusCodes.BAD_REQUEST).end();
+      }
 
-    return res.status(StatusCodes.OK).json(results);
-  });
+      return res.status(StatusCodes.OK).json(results);
+    });
+  }
 };
 
-function ensureAuthorization(req) {
-  let receivedJwt = req.headers['authorization'];
-  console.log('received jwt : ', receivedJwt);
+function ensureAuthorization(req, res) {
+  try {
+    let receivedJwt = req.headers['authorization'];
+    console.log('received jwt : ', receivedJwt);
 
-  let decodedJwt = jwt.verify(receivedJwt, process.env.PRIVATE_KEY);
-  console.log(decodedJwt);
+    let decodedJwt = jwt.verify(receivedJwt, process.env.PRIVATE_KEY);
+    console.log(decodedJwt);
 
-  return decodedJwt;
+    return decodedJwt;
+  } catch (err) {
+    console.log(err.name);
+    console.log(err.message);
+
+    return err;
+  }
 }
 
 module.exports = {
